pyproject.toml
[project]
name = "small-model-experiments-ib"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "llama-stack>=0.2.10.1",
    "llama-stack-client>=0.2.10",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.6",
    # Max Tool Experiment specific dependencies
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.2.0",
    "fastmcp>=0.1.0",
    "requests>=2.25.0",
    "mcp>=1.12.2",
    "openai>=1.97.1",
    "backoff>=2.2.1",
    "tqdm>=4.67.1",
    "torch>=2.8.0",
    "transformers>=4.55.4",
    "accelerate>=1.10.1",
    "pyyaml>=6.0.2",
]
